Update NaturalSplineCurve.h
Fixed gcc 12.1.1 compiler warnings (signed/unsigned integer comparisons).

+++ GTE/Mathematics/NaturalSplineCurve.h
@@ -3,7 +3,7 @@
 // Distributed under the Boost Software License, Version 1.0.
 // https://www.boost.org/LICENSE_1_0.txt
 // https://www.geometrictools.com/License/Boost/LICENSE_1_0.txt
-// Version: 6.0.2022.06.04
+// Version: 6.0.2022.06.07
 
 #pragma once
 
@@ -12,6 +12,10 @@
 
 namespace gte
 {
+    // NOTE: This class is now deprecated and will not be ported to GTL.
+    // Use instead the new class NaturalCubicSpline. There is also an
+    // extension of the idea in the new class NaturalQuinticSpline.
+
     template <int32_t N, typename Real>
     class NaturalSplineCurve : public ParametricCurve<N, Real>
     {
@@ -387,7 +391,7 @@ namespace gte
             }
             else
             {
-                for (int32_t i = 0, ip1 = 1; i < mNumSegments; i = ip1++)
+                for (size_t i = 0, ip1 = 1; i < mNumSegments; i = ip1++)
                 {
                     if (t < this->mTime[ip1])
                     {
